pyproject.toml
[tool.poetry]
name = "easyclock"
version = "0.1.0"
description = "A easy desktop clock app"
authors = ["ymy139 <ymy213456@163.com>"]
license = "AGPLv3"
repository = "https://github.com/ymy139/EasyClock"
readme = ["README.md", "LICENSE"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pyqt6 = "^6.7.0"
requests = "^2.31.0"
lunardate = "^0.2.2"
pyqt6-fluent-widgets = "^1.5.5"
pyinstaller = "^6.6.0"
winrt-windows-media-control = "^2.0.1"
winrt-windows-foundation = "^2.0.1"
winrt-windows-storage = "^2.0.1"
winrt-windows-storage-streams = "^2.0.1"
winrt-windows-foundation-collections = "^2.0.1"

[[tool.poetry.source]]
name = "mirrors_tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "mirrors_aliyun"
url = "http://mirrors.aliyun.com/pypi/simple/"
priority = "supplemental"


[[tool.poetry.source]]
name = "mirrors_bfsu"
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
